@@ -19,7 +19,7 @@ return [
 
     'allowed_methods' => ['*'],
 
-    'allowed_origins' => ['http://localhost:8000/prospects', 'http://127.0.0.1:8000/prospects'],
+    'allowed_origins' => ['http://localhost:8000/prospects', 'http://127.0.0.1:8000/prospects', 'http://127.0.0.1:8000', 'http://127.0.0.1:8000/api/prospects', 'http://localhost:8000/api/prospects', 'http://localhost:8000/api/prospects/filter', 'http://localhost:8000/api/*', 'http://127.0.0.1:8000/api/prospects/filter'],
 
     'allowed_origins_patterns' => [],
 
